requirements.txt
requests
pandas
pyarrow
configparser
python-dotenv